# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/issues:
    get:
      tags:
        - Issues
      summary: List Issues
      description: Get a paginated list of Issues with optional filtering and sorting
      operationId: listIssues
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Issues
      summary: Create Issue
      description: Create a new Issue
      operationId: createIssue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/service_level_agreements:
    get:
      tags:
        - ServiceLevelAgreements
      summary: List ServiceLevelAgreements
      description: Get a paginated list of ServiceLevelAgreements with optional filtering and sorting
      operationId: listServiceLevelAgreements
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceLevelAgreementList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ServiceLevelAgreements
      summary: Create ServiceLevelAgreement
      description: Create a new ServiceLevelAgreement
      operationId: createServiceLevelAgreement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceLevelAgreementRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceLevelAgreement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Issue:
      type: object
      required:
        - id
        - subject
        - priority
        - status
        - agreement_status
        - opened_at
        - response_breached
        - total_paused_mins
        - metadata
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          nullable: true
          type: string
          format: uuid
        subject:
          type: string
          maxLength: 200
        description:
          nullable: true
          type: string
          maxLength: 4000
        priority:
          type: string
        sla_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        agreement_status:
          type: string
        opened_at:
          type: string
          format: date-time
        response_by:
          nullable: true
          type: string
          format: date-time
        resolution_by:
          nullable: true
          type: string
          format: date-time
        first_responded_at:
          nullable: true
          type: string
          format: date-time
        response_breached:
          type: boolean
        resolved_at:
          nullable: true
          type: string
          format: date-time
        paused_at:
          nullable: true
          type: string
          format: date-time
        total_paused_mins:
          type: integer
          format: int32
        escalated_project_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateIssueRequest:
      type: object
      required:
        - subject
        - priority
        - status
        - agreement_status
        - opened_at
        - response_breached
        - total_paused_mins
        - metadata
      properties:
        customer_id:
          nullable: true
          type: string
          format: uuid
        subject:
          type: string
          maxLength: 200
        description:
          nullable: true
          type: string
          maxLength: 4000
        priority:
          type: string
        sla_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        agreement_status:
          type: string
        opened_at:
          type: string
          format: date-time
        response_by:
          nullable: true
          type: string
          format: date-time
        resolution_by:
          nullable: true
          type: string
          format: date-time
        first_responded_at:
          nullable: true
          type: string
          format: date-time
        response_breached:
          type: boolean
        resolved_at:
          nullable: true
          type: string
          format: date-time
        paused_at:
          nullable: true
          type: string
          format: date-time
        total_paused_mins:
          type: integer
          format: int32
        escalated_project_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    IssueList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Issue'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ServiceLevelAgreement:
      type: object
      required:
        - id
        - name
        - is_default
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 140
        is_default:
          type: boolean
        status:
          type: string
        metadata:
          type: object
...
    CreateServiceLevelAgreementRequest:
      type: object
      required:
        - name
        - is_default
        - status
        - metadata
      properties:
        name:
          type: string
          maxLength: 140
        is_default:
          type: boolean
        status:
          type: string
        metadata:
          type: object
...
    ServiceLevelAgreementList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ServiceLevelAgreement'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []