# src/domain/entity/service_level_agreement.rs
use chrono::{DateTime, Utc};
use serde::{Deserialize, Serialize};
use sqlx::FromRow;
use uuid::Uuid;

use super::AuditMetadata;
use super::ServiceLevelAgreementStatus;

/// Strongly-typed ID for ServiceLevelAgreement
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
#[serde(transparent)]
pub struct ServiceLevelAgreementId(pub Uuid);

impl ServiceLevelAgreementId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }
    pub fn generate() -> Self {
        Self(Uuid::new_v4())
    }
    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl std::fmt::Display for ServiceLevelAgreementId {
    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
        write!(f, "{}", self.0)
    }
}

impl std::str::FromStr for ServiceLevelAgreementId {
    type Err = uuid::Error;
    fn from_str(s: &str) -> Result<Self, Self::Err> {
        Ok(Self(Uuid::parse_str(s)?))
    }
}

impl From<Uuid> for ServiceLevelAgreementId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<ServiceLevelAgreementId> for Uuid {
    fn from(id: ServiceLevelAgreementId) -> Self {
        id.0
    }
}

impl AsRef<Uuid> for ServiceLevelAgreementId {
    fn as_ref(&self) -> &Uuid {
        &self.0
    }
}

impl std::ops::Deref for ServiceLevelAgreementId {
    type Target = Uuid;
    fn deref(&self) -> &Self::Target {
        &self.0
    }
}

#[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
pub struct ServiceLevelAgreement {
    pub id: Uuid,
    pub name: String,
    pub is_default: bool,
    pub status: ServiceLevelAgreementStatus,
    #[serde(default)]
    #[sqlx(json)]
    pub metadata: AuditMetadata,
}

impl ServiceLevelAgreement {
    /// Create a builder for ServiceLevelAgreement
    pub fn builder() -> ServiceLevelAgreementBuilder {
        <ServiceLevelAgreementBuilder as Default>::default()
    }

    /// Create a new ServiceLevelAgreement with required fields
    pub fn new(name: String, is_default: bool, status: ServiceLevelAgreementStatus) -> Self {
        Self {
            id: Uuid::new_v4(),
            name,
            is_default,
            status,
            metadata: AuditMetadata::default(),
        }
    }

    /// Get the entity's unique identifier
    pub fn id(&self) -> &Uuid {
        &self.id
    }

    /// Get a strongly-typed ID for this entity
    pub fn typed_id(&self) -> ServiceLevelAgreementId {
        ServiceLevelAgreementId(self.id)
    }

    /// Get when this entity was created
    pub fn created_at(&self) -> Option<&DateTime<Utc>> {
        self.metadata.created_at.as_ref()
    }

    /// Get when this entity was last updated
    pub fn updated_at(&self) -> Option<&DateTime<Utc>> {
        self.metadata.updated_at.as_ref()
    }

    /// Check if this entity is soft deleted
    pub fn is_deleted(&self) -> bool {
        self.metadata.deleted_at.is_some()
    }

    /// Check if this entity is active (not deleted)
    pub fn is_active(&self) -> bool {
        self.metadata.deleted_at.is_none()
    }

    /// Get when this entity was deleted
    pub fn deleted_at(&self) -> Option<&DateTime<Utc>> {
        self.metadata.deleted_at.as_ref()
    }

    /// Get who created this entity
    pub fn created_by(&self) -> Option<&Uuid> {
        self.metadata.created_by.as_ref()
    }

    /// Get who last updated this entity
    pub fn updated_by(&self) -> Option<&Uuid> {
        self.metadata.updated_by.as_ref()
    }

    /// Get who deleted this entity
    pub fn deleted_by(&self) -> Option<&Uuid> {
        self.metadata.deleted_by.as_ref()
    }

    /// Get the current status
    pub fn status(&self) -> &ServiceLevelAgreementStatus {
        &self.status
    }

    // ==========================================================
    // Partial Update
    // ==========================================================

    /// Apply partial updates from a map of field name to JSON value
    pub fn apply_patch(&mut self, fields: std::collections::HashMap<String, serde_json::Value>) {
        for (key, value) in fields {
            match key.as_str() {
                "name" => {
                    if let Ok(v) = serde_json::from_value(value) {
                        self.name = v;
                    }
                }
                "is_default" => {
                    if let Ok(v) = serde_json::from_value(value) {
                        self.is_default = v;
                    }
                }
                "status" => {
                    if let Ok(v) = serde_json::from_value(value) {
                        self.status = v;
                    }
                }
                _ => {} // ignore unknown fields
            }
        }
    }

    // <<< CUSTOM METHODS START >>>
    // <<< CUSTOM METHODS END >>>
}

impl super::Entity for ServiceLevelAgreement {
    type Id = Uuid;

    fn entity_id(&self) -> &Self::Id {
        &self.id
    }

    fn entity_type() -> &'static str {
        "ServiceLevelAgreement"
    }
}

impl backbone_core::PersistentEntity for ServiceLevelAgreement {
    fn entity_id(&self) -> String {
        self.id.to_string()
    }
    fn set_entity_id(&mut self, id: String) {
        if let Ok(uuid) = uuid::Uuid::parse_str(&id) {
            self.id = uuid;
        }
    }
    fn created_at(&self) -> Option<chrono::DateTime<chrono::Utc>> {
        self.metadata.created_at
    }
    fn set_created_at(&mut self, ts: chrono::DateTime<chrono::Utc>) {
        self.metadata.created_at = Some(ts);
    }
    fn updated_at(&self) -> Option<chrono::DateTime<chrono::Utc>> {
        self.metadata.updated_at
    }
    fn set_updated_at(&mut self, ts: chrono::DateTime<chrono::Utc>) {
        self.metadata.updated_at = Some(ts);
    }
    fn deleted_at(&self) -> Option<chrono::DateTime<chrono::Utc>> {
        self.metadata.deleted_at
    }
    fn set_deleted_at(&mut self, ts: Option<chrono::DateTime<chrono::Utc>>) {
        self.metadata.deleted_at = ts;
    }
}

impl backbone_orm::EntityRepoMeta for ServiceLevelAgreement {
    fn column_types() -> std::collections::HashMap<String, String> {
        let mut m = std::collections::HashMap::new();
        m.insert("id".to_string(), "uuid".to_string());
        m.insert(
            "status".to_string(),
            "service_level_agreement_status".to_string(),
        );
        m
    }
    fn search_fields() -> &'static [&'static str] {
        &["name"]
    }
}

/// Builder for ServiceLevelAgreement entity
///
/// Provides a fluent API for constructing ServiceLevelAgreement instances.
/// System fields (id, metadata, timestamps) are auto-initialized.
#[derive(Debug, Clone, Default)]
pub struct ServiceLevelAgreementBuilder {
    name: Option<String>,
    is_default: Option<bool>,
    status: Option<ServiceLevelAgreementStatus>,
}

impl ServiceLevelAgreementBuilder {
    /// Set the name field (required)
    pub fn name(mut self, value: String) -> Self {
        self.name = Some(value);
        self
    }

    /// Set the is_default field (default: `false`)
    pub fn is_default(mut self, value: bool) -> Self {
        self.is_default = Some(value);
        self
    }

    /// Set the status field (default: `ServiceLevelAgreementStatus::default()`)
    pub fn status(mut self, value: ServiceLevelAgreementStatus) -> Self {
        self.status = Some(value);
        self
    }

    /// Build the ServiceLevelAgreement entity
    ///
    /// Returns Err if any required field without a default is missing.
    pub fn build(self) -> Result<ServiceLevelAgreement, String> {
        let name = self.name.ok_or_else(|| "name is required".to_string())?;

        Ok(ServiceLevelAgreement {
            id: Uuid::new_v4(),
            name,
            is_default: self.is_default.unwrap_or(false),
            status: self.status.unwrap_or_default(),
            metadata: AuditMetadata::default(),
        })
    }
}